+++ cas_work_report/models/__init__.py
@@ -1,0 +1,2 @@
+from . import work_report
+from . import export_wizard